backend/requirements/base.txt
Django>=4.2,<5.0
djangorestframework>=3.14
django-cors-headers>=4.0
django-filter>=23.0
psycopg2-binary>=2.9
celery>=5.3
redis>=5.0
pillow>=10.0
drf-spectacular>=0.27
django-environ>=0.11
djangorestframework-simplejwt>=5.3
